Add files via upload
With this code my goal was to compile data with python. Then the user requests a data set (in this case they type in the day of the week in the console) and the program shows them that specific data set (in this week a class schedule for that day). I was learning about tuples and functions and trying to get a greater understanding on while loops and if statements. I was also experimenting with user interaction with consol input. This is the result. 2022/12/18